requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
python-dotenv
python-telegram-bot
numpy
requests
google-cloud-pubsub
ebaysdk

